+++ src/RegistrationError.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Bakame\Cron;
+
+use LogicException;
+
+final class RegistrationError extends LogicException implements CronError
+{
+}